fix: FilterProduct, SeachCity, AuthChecker
