+++ src/components/FilterProduct/FilterProduct.scss
@@ -33,7 +33,7 @@
         gap: 5px;
         margin-bottom: 25px;
 
-        &>input {
+        input {
             background: #F4F4F4;
             display: block;
             font-size: 18px;
