+++ src/components/MainPageNewAnnouncements/MainPageNewAnnouncements.jsx
@@ -13,8 +13,6 @@ export default function MainPageNewAnnouncements() {
     });
   }, [setAllProduct]);
 
-  console.log(allProduct);
-
   return (
     <div className="NewAnnouncements">
       <h2>Найновіші оголошення</h2>
